src/backend/requirements.txt
# Web Framework and Server
fastapi==0.115.12

# LlamaIndex Components
llama-index-core==0.12.37
llama-index-embeddings-openai==0.3.1
llama-index-readers-web==0.4.1
llama-index-vector-stores-qdrant==0.6.0

# Language Models
openai==1.81.0

# Data Validation
pydantic==2.11.4

# Document Processing
PyMuPDF==1.26.0

# Configuration
python-dotenv==1.1.0
python-multipart==0.0.20

# Vector Database
qdrant-client==1.14.2
tiktoken==0.9.0
trafilatura==2.0.0

# Miscellaneous
uvicorn==0.34.2
